photo_album_src/requirements.txt
flask
boto3
werkzeug
pymongo
Pillow
uuid
